requirements.txt
numpy
opencv-python
onnxruntime
scipy
pillow
requests
tqdm
pyyaml
